django/requirements.txt
django==2.1
psycopg2-binary
django-crispy-forms==1.7.0
django-haystack
django-haystack-es
elasticsearch>=5,<6
coverage===4.4.2
simplejson
requests
